--- backend/trafficrules/src/main/java/com/coffeejug/trafficrules/db/User.java
@@ -1,46 +1,0 @@
-package com.coffeejug.trafficrules.db;
-
-import org.hibernate.annotations.GenericGenerator;
-
-import javax.persistence.*;
-import java.util.UUID;
-
-@Entity
-@Table(name = "users")
-public class User {
-
-    @Id
-    @GeneratedValue(generator = "uuid2")
-    @GenericGenerator(name = "uuid2", strategy = "uuid2")
-    @Column(columnDefinition = "BINARY(16)")
-    private UUID uuid;
-
-    @Column(columnDefinition = "INT NOT NULL DEFAULT '0'")
-    private int levelsCompleted;
-
-    private String name;
-
-    public UUID getUuid() {
-        return uuid;
-    }
-
-    public void setUuid(UUID uuid) {
-        this.uuid = uuid;
-    }
-
-    public int getLevelsCompleted() {
-        return levelsCompleted;
-    }
-
-    public void setLevelsCompleted(int levelsCompleted) {
-        this.levelsCompleted = levelsCompleted;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
